# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Visitor Management System API
  description: Comprehensive API for managing visitors, visits, and organizational access control
  version: 1.0.0
  contact:
    name: VMS API Support
    email: support@vms.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.vms.example.com/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        org_slug:
          type: string
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/Employee'
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # Core Models
...
    Employee:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/EmployeeRole'
        is_host:
          type: boolean
        department:
          type: string
        job_title:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    EmployeeRole:
      type: string
      enum: [ADMIN, RECEPTIONIST, SECURITY, MANAGER]
    EmployeeStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...